Final Project 99%

+++ src/Playable.java
@@ -2,6 +2,6 @@ public interface Playable {
     void play();
     void pause();
     void stop();
-    int getDurationSrc();
+    long getDurationSrc();
 }
